Skeleton for Integration Tests DONE
