+++ Session.vim
@@ -13,14 +13,10 @@ if &shortmess =~ 'A'
 else
   set shortmess=aoO
 endif
-badd +14 src/main.rs
-badd +94 src/profanity.rs
-badd +113 src/routes/authentication.rs
 argglobal
 %argdel
-edit src/routes/authentication.rs
 argglobal
-balt src/profanity.rs
+enew
 setlocal fdm=manual
 setlocal fde=0
 setlocal fmr={{{,}}}
@@ -29,42 +25,6 @@ setlocal fdl=99
 setlocal fml=1
 setlocal fdn=20
 setlocal fen
-silent! normal! zE
-1,5fold
-7,9fold
-16,19fold
-22,25fold
-13,26fold
-29,33fold
-42,43fold
-42,44fold
-41,44fold
-45,47fold
-40,48fold
-49,50fold
-38,52fold
-37,55fold
-36,56fold
-59,61fold
-69,71fold
-64,80fold
-85,90fold
-83,94fold
-99,101fold
-98,104fold
-98,105fold
-97,106fold
-108,110fold
-116,127fold
-112,128fold
-let &fdl = &fdl
-let s:l = 113 - ((25 * winheight(0) + 25) / 51)
-if s:l < 1 | let s:l = 1 | endif
-keepjumps exe s:l
-normal! zt
-keepjumps 113
-normal! 023|
-if exists(':tcd') == 2 | tcd ~/DEV/RUST_WEB_DEVELOPMENT/CH_10/rust-web-dev-config | endif
 tabnext 1
 if exists('s:wipebuf') && len(win_findbuf(s:wipebuf)) == 0 && getbufvar(s:wipebuf, '&buftype') isnot# 'terminal'
   silent exe 'bwipe ' . s:wipebuf
